data-processing-service/requirements.txt
flask==2.3.3
flask-restful==0.3.10
pyspark==3.5.0
kafka-python==2.0.2
redis==5.0.1
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.24.3
pytest==7.4.2
black==23.7.0
flake8==6.1.0
tensorflow==2.13.0
scikit-learn==1.3.0
transformers==4.31.0
torch==2.0.1
pytz==2023.3
nltk==3.8.1 